@@ -1,77 +1,50 @@
-import axios from 'axios'
-import { Context } from 'koa'
-import { getOrCreateUser } from '@/models/user'
 import { Body, Controller, Ctx, Post } from 'amala'
-import Facebook = require('facebook-node-sdk')
-import {
-  TelegramLoginPayload,
-  verifyTelegramPayload,
-} from '@/helpers/verifyTelegramPayload'
+import { Context } from 'koa'
+import { forbidden } from '@hapi/boom'
+import { getOrCreateUser } from '@/models/user'
+import { verifyTelegramPayload } from '@/helpers/verifyTelegramPayload'
+import FacebookLogin from '@/validators/FacebookLogin'
+import GoogleLogin from '@/validators/GoogleLogin'
+import TelegramLogin from '@/validators/TelegramLogin'
+import getFBUser from '@/helpers/getFBUser'
+import getGoogleUser from '@/helpers/getGoogleUser'
 
 @Controller('/login')
 export default class LoginController {
   @Post('/facebook')
-  async facebook(@Body('accessToken') accessToken: string) {
-    const fbProfile: any = await getFBUser(accessToken)
+  async facebook(@Body({ required: true }) { accessToken }: FacebookLogin) {
+    const { name, email, id } = await getFBUser(accessToken)
     const user = await getOrCreateUser({
-      name: fbProfile.name,
-
-      email: fbProfile.email,
-      facebookId: fbProfile.id,
+      name,
+      email,
+      facebookId: id,
     })
-    return user.strippedAndFilled(true)
+    return user.strippedAndFilled({ withExtra: true })
   }
 
   @Post('/telegram')
-  async telegram(@Ctx() ctx: Context, @Body() data: TelegramLoginPayload) {
-    // verify the data
-    if (!verifyTelegramPayload(data)) {
-      return ctx.throw(403)
+  async telegram(
+    @Ctx() ctx: Context,
+    @Body({ required: true }) body: TelegramLogin,
+    @Body({ required: true }) { first_name, last_name, id }: TelegramLogin
+  ) {
+    if (!verifyTelegramPayload(body)) {
+      return ctx.throw(forbidden())
     }
-
     const user = await getOrCreateUser({
-      name: `${data.first_name}${data.last_name ? ` ${data.last_name}` : ''}`,
-      telegramId: data.id,
+      name: `${first_name}${last_name ? ` ${last_name}` : ''}`,
+      telegramId: id,
     })
-    return user.strippedAndFilled(true)
+    return user.strippedAndFilled({ withExtra: true })
   }
 
   @Post('/google')
-  async google(@Body('accessToken') accessToken: string) {
-    const userData: any =
-      process.env.TESTING === 'true'
-        ? testingGoogleMock()
-        : (
-            await axios(
-              `https://www.googleapis.com/oauth2/v3/tokeninfo?id_token=${accessToken}`
-            )
-          ).data
-
+  async google(@Body({ required: true }) { accessToken }: GoogleLogin) {
+    const userData = await getGoogleUser(accessToken)
     const user = await getOrCreateUser({
       name: userData.name,
-
       email: userData.email,
     })
-    return user.strippedAndFilled(true)
-  }
-}
-
-function getFBUser(accessToken: string) {
-  return new Promise((res, rej) => {
-    const fb = new Facebook({
-      appID: process.env.FACEBOOK_APP_ID,
-      secret: process.env.FACEBOOK_APP_SECRET,
-    })
-    fb.setAccessToken(accessToken)
-    fb.api('/me?fields=name,email,id', (err, user) => {
-      return err ? rej(err) : res(user)
-    })
-  })
-}
-
-function testingGoogleMock() {
-  return {
-    name: 'Alexander Brennenburg',
-    email: 'alexanderrennenburg@gmail.com',
+    return user.strippedAndFilled({ withExtra: true })
   }
 }
