@@ -1,12 +1,12 @@
 import 'reflect-metadata'
 import * as Koa from 'koa'
-import * as bodyParser from 'koa-bodyparser'
-import { bootstrapControllers } from 'amala'
-import * as cors from '@koa/cors'
 import * as Router from 'koa-router'
+import * as bodyParser from 'koa-bodyparser'
+import * as cors from '@koa/cors'
+import { bootstrapControllers } from 'amala'
 
-export const app = new Koa()
-;(async () => {
+const app = new Koa()
+void (async () => {
   try {
     const router = new Router()
     await bootstrapControllers({
@@ -24,3 +24,5 @@ export const app = new Koa()
     console.log('Koa app starting error: ', err)
   }
 })()
+
+export default app
