@@ -1,10 +1,10 @@
 import * as jwt from 'jsonwebtoken'
 
-const secret = process.env.JWT
+const getSecret = () => process.env.JWT
 
-export function sign(payload: object) {
+export function sign(payload: Record<string, unknown>) {
   return new Promise((res, rej) => {
-    jwt.sign(payload, secret, undefined, (err, token) => {
+    jwt.sign(payload, getSecret(), undefined, (err, token) => {
       return err ? rej(err) : res(token)
     })
   })
@@ -12,7 +12,7 @@ export function sign(payload: object) {
 
 export function verify(token: string) {
   return new Promise((res, rej) => {
-    jwt.verify(token, secret, undefined, (err, payload) => {
+    jwt.verify(token, getSecret(), undefined, (err, payload) => {
       return err ? rej(err) : res(payload)
     })
   })
