@@ -1,5 +1,5 @@
 // Setup typegoose
-import { setGlobalOptions, Severity } from '@typegoose/typegoose'
+import { Severity, setGlobalOptions } from '@typegoose/typegoose'
 setGlobalOptions({
   options: {
     allowMixed: Severity.ALLOW,
@@ -8,14 +8,15 @@ setGlobalOptions({
 import 'module-alias/register'
 import * as dotenv from 'dotenv'
 dotenv.config({ path: `${__dirname}/../.env` })
-import { app } from '@/app'
 import { runMongo } from '@/models/index'
+import app from '@/app'
 
 // Run mongo
-runMongo().then(() => {
+void runMongo().then(() => {
   console.log('Mongo connected')
 })
 // Start rest
-app.listen(1337).on('listening', () => {
-  console.log('HTTP is listening on 1337')
+const port = process.env.PORT || 1337
+app.listen(port).on('listening', () => {
+  console.log(`HTTP is listening on ${port}`)
 })
